my personal details with code.
Welcome to my personal portfolio hosted on GitHub! This repository serves as a window into my journey as a developer, featuring an array of projects, contributions, and experiments. From web applications to machine learning models, you'll find a diverse range of endeavors reflecting my passion for technology and innovation. Dive into the code, explore my learning process, and witness the evolution of my skills firsthand. Join me on this exciting adventure as we embrace the endless possibilities of software development. Explore my GitHub portfolio today and embark on a journey of exploration and discovery!